+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/impl/EnumType.java
@@ -6,6 +6,7 @@ package com.graphql_java_generator.plugin.language.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.graphql_java_generator.plugin.DocumentParser;
 import com.graphql_java_generator.plugin.conf.CommonConfiguration;
 import com.graphql_java_generator.plugin.conf.GenerateCodeCommonConfiguration;
 import com.graphql_java_generator.plugin.language.EnumValue;
@@ -33,9 +34,12 @@ public class EnumType extends AbstractType {
 	 * @param configuration
 	 *            The current plugin configuration, which is accessible through an interface that extends
 	 *            {@link CommonConfiguration}
+	 * @param documentParser
+	 *            The {@link DocumentParser} that has parsed the schema, and so that contains the whole schema
+	 *            definition
 	 */
-	public EnumType(String name, CommonConfiguration configuration) {
-		super(name, GraphQlType.ENUM, configuration);
+	public EnumType(String name, CommonConfiguration configuration, DocumentParser documentParser) {
+		super(name, GraphQlType.ENUM, configuration, documentParser);
 	}
 
 	/**