+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/impl/InterfaceType.java
@@ -6,6 +6,7 @@ package com.graphql_java_generator.plugin.language.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.graphql_java_generator.plugin.DocumentParser;
 import com.graphql_java_generator.plugin.conf.CommonConfiguration;
 
 import lombok.Data;
@@ -64,9 +65,12 @@ public class InterfaceType extends ObjectType {
 	 * @param configuration
 	 *            The current plugin configuration, which is accessible through an interface that extends
 	 *            {@link CommonConfiguration}
+	 * @param documentParser
+	 *            The {@link DocumentParser} that has parsed the schema, and so that contains the whole schema
+	 *            definition
 	 */
-	public InterfaceType(String name, CommonConfiguration configuration) {
-		super(name, GraphQlType.INTERFACE, configuration);
+	public InterfaceType(String name, CommonConfiguration configuration, DocumentParser documentParser) {
+		super(name, GraphQlType.INTERFACE, configuration, documentParser);
 	}
 
 	@Override