+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/impl/ScalarType.java
@@ -3,6 +3,7 @@
  */
 package com.graphql_java_generator.plugin.language.impl;
 
+import com.graphql_java_generator.plugin.DocumentParser;
 import com.graphql_java_generator.plugin.conf.CommonConfiguration;
 import com.graphql_java_generator.plugin.conf.GraphQLConfiguration;
 import com.graphql_java_generator.plugin.language.Field;
@@ -32,9 +33,13 @@ public class ScalarType extends AbstractType {
 	 *            The simple name (without the package) of the java class that will hold this value
 	 * @param configuration
 	 *            The current {@link GraphQLConfiguration}
+	 * @param documentParser
+	 *            The {@link DocumentParser} that has parsed the schema, and so that contains the whole schema
+	 *            definition
 	 */
-	public ScalarType(String name, String packageName, String classSimpleName, CommonConfiguration configuration) {
-		super(name, GraphQlType.SCALAR, configuration);
+	public ScalarType(String name, String packageName, String classSimpleName, CommonConfiguration configuration,
+			DocumentParser documentParser) {
+		super(name, GraphQlType.SCALAR, configuration, documentParser);
 		this.packageName = packageName;
 		this.classSimpleName = classSimpleName;
 	}