+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/language/impl/AbstractTypeTest.java
@@ -18,7 +18,7 @@ class AbstractTypeTest {
 	@Test
 	void testAddImportClassOfQ() {
 		// Preparation
-		AbstractType type = new ObjectType("DateFormat", pluginConfiguration);
+		AbstractType type = new ObjectType("DateFormat", pluginConfiguration, null);
 		assertEquals(0, type.getImports().size());
 
 		// Add of a class from the same package
@@ -52,7 +52,7 @@ class AbstractTypeTest {
 	@Test
 	void testAddImportStringString() {
 		// Preparation
-		AbstractType type = new ObjectType("DateFormat", pluginConfiguration);
+		AbstractType type = new ObjectType("DateFormat", pluginConfiguration, null);
 		assertEquals(0, type.getImports().size());
 
 		// Add of a class from the same package